# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "{{.Description}}",
        "title": "{{.Title}}",
        "contact": {},
        "license": {},
        "version": "{{.Version}}"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
...
        "/merchant/auditstatus": {
            "get": {
                "description": "获取承兑商账号审核状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "承兑商APP API"
                ],
                "summary": "获取承兑商账号审核状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "uid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.GetAuditStatusRet"
                        }
                    }
                }
            }
        },
...
        "/merchant/profile": {
            "get": {
                "description": "获取承兑商个人信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "承兑商APP API"
                ],
                "summary": "获取承兑商个人信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "uid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.GetProfileRet"
                        }
                    }
                }
            }
        },
...
        "/merchant/settings/identify": {
            "get": {
                "description": "获取承兑商订单推送模式和开关",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "承兑商APP API"
                ],
                "summary": "获取承兑商订单推送模式和开关",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "uid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.GetIdentifyRet"
                        }
                    }
                }
            },
            "put": {
                "description": "承兑商设置自己的认证信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "承兑商APP API"
                ],
                "summary": "承兑商设置自己的认证信息",
                "parameters": [
                    {
                        "description": "用户id",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.SetIdentifyArg"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.SetIdentifyRet"
                        }
                    }
                }
            }
        },
        "/merchant/settings/nickname": {
            "put": {
                "description": "设置承兑商昵称",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "承兑商APP API"
                ],
                "summary": "设置承兑商昵称",
                "parameters": [
                    {
                        "description": "用户id",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.SetNickNameArg"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.SetNickNameRet"
                        }
                    }
                }
            }
        },
...
        "/merchant/settings/workmode": {
            "get": {
                "description": "获取承兑商订单推送模式和开关",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "承兑商APP API"
                ],
                "summary": "获取承兑商订单推送模式和开关",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "uid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.GetWorkModeRet"
                        }
                    }
                }
            },
            "put": {
                "description": "承兑商设置订单推送模式和开关",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "承兑商APP API"
                ],
                "summary": "承兑商设置订单推送模式和开关",
                "parameters": [
                    {
                        "description": "用户id",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.SetWorkModeArg"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.SetWorkModeRet"
                        }
                    }
                }
            }
        },
        "/merchants": {
            "get": {
                "description": "坐席获取承兑商列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台 API"
                ],
                "summary": "获取承兑商列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页数",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "承兑商状态",
                        "name": "user_status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "筛选开始时间",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "筛选截止时间",
                        "name": "stop_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "筛选字段",
                        "name": "time_field",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索值",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功（status为success）失败（status为fail）都会返回200",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.MerchantRet"
                        }
                    }
                }
            }
        },
        "/merchants/{uid}/approve": {
            "put": {
                "description": "审核承兑商",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台 API"
                ],
                "summary": "审核",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "充值",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.ApproveArgs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功（status为success）失败（status为fail）都会返回200",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.ApproveRet"
                        }
                    }
                }
            }
        },
        "/merchants/{uid}/asset": {
            "put": {
                "description": "给承兑商充值",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台 API"
                ],
                "summary": "充值",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "充值",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.RechargeArgs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功（status为success）失败（status为fail）都会返回200",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.RechargeRet"
                        }
                    }
                }
            }
        },
...
        "/merchants/{uid}/freeze": {
            "put": {
                "description": "审核承冻结或者解冻",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台 API"
                ],
                "summary": "冻结",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "冻结操作",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.FreezeArgs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功（status为success）失败（status为fail）都会返回200",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.MerchantRet"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.ApproveArgs": {
            "type": "object",
            "required": [
                "count",
                "currency",
                "operation"
            ],
            "properties": {
                "count": {
                    "type": "string",
                    "example": "200"
                },
                "currency": {
                    "type": "string",
                    "example": "BTUSD"
                },
                "operation": {
                    "description": "审核操作 1：通过 0：不通过",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "response.ApproveRet": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "entity": {
                    "type": "object",
                    "properties": {
                        "status": {
                            "type": "integer"
                        },
                        "uid": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                },
                "err_code": {
                    "description": "err_code仅在失败时设置",
                    "type": "integer"
                },
                "err_msg": {
                    "description": "err_msg仅在失败时设置",
                    "type": "string",
                    "example": "由于xx原因，导致操作失败"
                },
                "status": {
                    "description": "status可以为success或者fail",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "response.GetAuditStatusRet": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "entity": {
                    "type": "object",
                    "properties": {
                        "contact_phone": {
                            "description": "客服联系信息",
                            "type": "string",
                            "example": "13812341234"
                        },
                        "extra_message": {
                            "description": "额外的信息",
                            "type": "string",
                            "example": "您由于xx原因，未通过审核"
                        },
                        "uid": {
                            "description": "用户id",
                            "type": "integer"
                        },
                        "user_status": {
                            "description": "user_status可以为0/1/2，分别表示“正常/待审核/冻结”",
                            "type": "integer"
                        }
                    }
                },
                "err_code": {
                    "description": "err_code仅在失败时设置",
                    "type": "integer"
                },
                "err_msg": {
                    "description": "err_msg仅在失败时设置",
                    "type": "string",
                    "example": "由于xx原因，导致操作失败"
                },
                "status": {
                    "description": "status可以为success或者fail",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "response.GetIdentifyRet": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "entity": {
                    "type": "object",
                    "properties": {
                        "email": {
                            "type": "string",
                            "example": "xxx@xxx.com"
                        },
                        "idcard": {
                            "type": "string",
                            "example": "11088888888888888"
                        },
                        "phone": {
                            "type": "string"
                        },
                        "uid": {
                            "description": "用户id",
                            "type": "integer"
                        }
                    }
                },
                "err_code": {
                    "description": "err_code仅在失败时设置",
                    "type": "integer"
                },
                "err_msg": {
                    "description": "err_msg仅在失败时设置",
                    "type": "string",
                    "example": "由于xx原因，导致操作失败"
                },
                "status": {
                    "description": "status可以为success或者fail",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "response.GetProfileRet": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "entity": {
                    "type": "object",
                    "properties": {
                        "asset_frozen": {
                            "description": "当前承兑商被冻结的平台币数量",
                            "type": "string",
                            "example": "100"
                        },
                        "asset_symbol": {
                            "description": "平台币的符号",
                            "type": "string",
                            "example": "BTUSD"
                        },
                        "asset_total": {
                            "description": "当前承兑商所有的平台币余额（包含被冻结的平台币）",
                            "type": "string",
                            "example": "2000"
                        },
                        "nickname": {
                            "description": "用户昵称",
                            "type": "string",
                            "example": "老王"
                        },
                        "uid": {
                            "description": "用户id",
                            "type": "integer"
                        }
                    }
                },
                "err_code": {
                    "description": "err_code仅在失败时设置",
                    "type": "integer"
                },
                "err_msg": {
                    "description": "err_msg仅在失败时设置",
                    "type": "string",
                    "example": "由于xx原因，导致操作失败"
                },
                "status": {
                    "description": "status可以为success或者fail",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "response.GetWorkModeRet": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "entity": {
                    "type": "object",
                    "properties": {
                        "accept": {
                            "description": "是否接单(1:开启，0:关闭)",
                            "type": "integer"
                        },
                        "auto": {
                            "description": "是否自动接单(1:开启，0:关闭)",
                            "type": "integer"
                        },
                        "uid": {
                            "description": "用户id",
                            "type": "integer"
                        }
                    }
                },
                "err_code": {
                    "description": "err_code仅在失败时设置",
                    "type": "integer"
                },
                "err_msg": {
                    "description": "err_msg仅在失败时设置",
                    "type": "string",
                    "example": "由于xx原因，导致操作失败"
                },
                "status": {
                    "description": "status可以为success或者fail",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "response.MerchantRet": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "entity": {
                    "type": "object",
                    "properties": {
                        "data": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Merchant"
                            }
                        }
                    }
                },
                "err_code": {
                    "description": "err_code仅在失败时设置",
                    "type": "integer"
                },
                "err_msg": {
                    "description": "err_msg仅在失败时设置",
                    "type": "string",
                    "example": "由于xx原因，导致操作失败"
                },
                "status": {
                    "description": "status可以为success或者fail",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "response.RechargeArgs": {
            "type": "object",
            "required": [
                "count",
                "currency"
            ],
            "properties": {
                "count": {
                    "type": "string",
                    "example": "200"
                },
                "currency": {
                    "type": "string",
                    "example": "BTUSD"
                }
            }
        },
        "response.RechargeRet": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "entity": {
                    "type": "object",
                    "properties": {
                        "balance": {
                            "type": "string"
                        }
                    }
                },
                "err_code": {
                    "description": "err_code仅在失败时设置",
                    "type": "integer"
                },
                "err_msg": {
                    "description": "err_msg仅在失败时设置",
                    "type": "string",
                    "example": "由于xx原因，导致操作失败"
                },
                "status": {
                    "description": "status可以为success或者fail",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "response.SetIdentifyArg": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "xxx@xxx.com"
                },
                "idcard": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string"
                },
                "uid": {
                    "description": "用户id",
                    "type": "integer"
                }
            }
        },
        "response.SetIdentifyRet": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "entity": {
                    "type": "object",
                    "properties": {
                        "uid": {
                            "description": "用户id",
                            "type": "integer"
                        }
                    }
                },
                "err_code": {
                    "description": "err_code仅在失败时设置",
                    "type": "integer"
                },
                "err_msg": {
                    "description": "err_msg仅在失败时设置",
                    "type": "string",
                    "example": "由于xx原因，导致操作失败"
                },
                "status": {
                    "description": "status可以为success或者fail",
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "response.SetNickNameArg": {
            "type": "object",
            "properties": {
                "nickname": {
                    "description": "想设置的新昵称",
                    "type": "string",
                    "example": "王老板"
                },
                "uid": {
                    "description": "用户id",
                    "type": "integer"
                }
            }
        },
        "response.SetNickNameRet": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "entity": {
                    "type": "object",
                    "properties": {
                        "uid": {
                            "description": "用户id",
                            "type": "integer"
                        }
                    }
                },
                "err_code": {
                    "description": "err_code仅在失败时设置",
                    "type": "integer"
                },
                "err_msg": {
                    "description": "err_msg仅在失败时设置",
                    "type": "string",
                    "example": "由于xx原因，导致操作失败"
                },
                "status": {
                    "description": "status可以为success或者fail",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "response.SetWorkModeArg": {
            "type": "object",
            "properties": {
                "accept": {
                    "description": "是否接单(1:开启，0:关闭)",
                    "type": "integer"
                },
                "auto": {
                    "description": "是否自动接单(1:开启，0:关闭)",
                    "type": "integer"
                },
                "uid": {
                    "description": "用户id",
                    "type": "integer"
                }
            }
        },
        "response.SetWorkModeRet": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "entity": {
                    "type": "object",
                    "properties": {
                        "uid": {
                            "description": "用户id",
                            "type": "integer"
                        }
                    }
                },
                "err_code": {
                    "description": "err_code仅在失败时设置",
                    "type": "integer"
                },
                "err_msg": {
                    "description": "err_msg仅在失败时设置",
                    "type": "string",
                    "example": "由于xx原因，导致操作失败"
                },
                "status": {
                    "description": "status可以为success或者fail",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
    }
}